Support setting unverified custom text answers
